multipleFiles

+++ multipleFiles/multiplefiles.h
@@ -4,7 +4,7 @@
 const float PARAMETER_RATIO = 1.3;
 const float PARAMETER_POWER = 0.625;
 const float PARAMETER_POWER2 = 0.25;
-double CalEquivDuct(double, double);
+float CalEquivDuct(double, double);
 
 
 #endif // !DUCTS_H_INCLUDED
